+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 next
 ----
 
+- fix infinite recursion when custom view type is itself (#4198)
 - respect firstDay setting when weekNumberCalculation set to ISO (#4734)
 - fix typo in Danish (#4708)
 - adjust typescript def for setExtendedProp (#4679)